Removed init and deactivate
I don't see a real use for it for now, since page gets reloaded every time anyways so what should happen during plugin init that can't happen on activation? Similar for deactivate.
For init, maybe later on two 'rounds' could be needed where the first looks at all the settings and other enabled plugins and changes behavior based on it, and a second round where stuff actually happens. But even then, it seems better that plugins don't rely too much on eacht other else being the same plugin to begin with.
Deactivate wasn't implemented anyways and can be implemented for real when it is actually needed.